+++ packages/cli/__tests__/zosfiles/__system__/download/ds/cli.files.download.ds.system.test.ts
@@ -109,6 +109,22 @@ describe("Download Data Set", () => {
             expect(response.stdout.toString()).toContain("Data set downloaded successfully.");
         });
 
+        it("should download data set in binary format", async () => {
+            const shellScript = path.join(__dirname, "__scripts__", "command", "command_download_data_set.sh");
+            const response = runCliScript(shellScript, TEST_ENVIRONMENT, [dsname, "--binary"]);
+            expect(response.stderr.toString()).toBe("");
+            expect(response.status).toBe(0);
+            expect(response.stdout.toString()).toContain("Data set downloaded successfully.");
+        });
+
+        it("should download data set in record format", async () => {
+            const shellScript = path.join(__dirname, "__scripts__", "command", "command_download_data_set.sh");
+            const response = runCliScript(shellScript, TEST_ENVIRONMENT, [dsname, "--record"]);
+            expect(response.stderr.toString()).toBe("");
+            expect(response.status).toBe(0);
+            expect(response.stdout.toString()).toContain("Data set downloaded successfully.");
+        });
+
         it("should download data set with response timeout", async () => {
             const shellScript = path.join(__dirname, "__scripts__", "command", "command_download_data_set.sh");
             const response = runCliScript(shellScript, TEST_ENVIRONMENT, [dsname, "--responseTimeout 5"]);