+++ packages/cli/src/zosfiles/upload/ftds/FileToDataSet.handler.ts
@@ -32,6 +32,7 @@ export default class FileToDataSetHandler extends ZosFilesBaseHandler {
             {
                 volume: commandParameters.arguments.volumeSerial,
                 binary: commandParameters.arguments.binary,
+                record: commandParameters.arguments.record,
                 encoding: commandParameters.arguments.encoding,
                 task,
                 responseTimeout: commandParameters.arguments.responseTimeout