+++ packages/cli/src/zosjobs/delete/job/Job.handler.ts
@@ -9,7 +9,7 @@
 *
 */
 
-import { IHandlerParameters } from "@zowe/imperative";
+import { IHandlerParameters, ImperativeError } from "@zowe/imperative";
 import { DeleteJobs, GetJobs, IJob } from "@zowe/zos-jobs-for-zowe-sdk";
 import { ZosmfBaseHandler } from "@zowe/zosmf-for-zowe-sdk";
 
@@ -43,9 +43,21 @@ export default class JobHandler extends ZosmfBaseHandler {
         // Get the job details
         const job: IJob = await GetJobs.getJob(this.mSession, jobid);
         // Delete the job
-        await DeleteJobs.deleteJobForJob(this.mSession, job, this.arguments.modifyVersion);
+        const response = await DeleteJobs.deleteJobForJob(this.mSession, job, this.arguments.modifyVersion);
 
-        const message: string = `Successfully deleted job ${job.jobname} (${jobid})`;
+        let message: string;
+        if (this.arguments.modifyVersion == null || this.arguments.modifyVersion === "1.0") {
+            message = `Successfully submitted request to delete job ${job.jobname} (${jobid})`;
+        } else if (this.arguments.modifyVersion === "2.0" && response?.status === "0") {
+            message = `Successfully deleted job ${job.jobname} (${jobid})`;
+        } else {
+            throw new ImperativeError({
+                msg: `Failed to delete job ${job.jobname} (${jobid})`,
+                additionalDetails: response?.message,
+                errorCode: response?.["internal-code"]
+            });
+        }
+
         // Print message to console
         this.console.log(message);